+++ diag_table.lua
@@ -5,7 +5,7 @@ function scene:create(event)
     local sceneGroup = self.view
 
     -- BACKGROUND
-    local bg = display.newImage("image/cutscene/cutscene_4.png")
+    local bg = display.newImage("image/cutscene/cutscene_5.png")
     bg.x = display.contentCenterX
     bg.y = display.contentCenterY
 
@@ -33,80 +33,11 @@ function scene:create(event)
     if (Data) then print(Data[1].dialogue) end
 
     local index = 0
-    local horror = "change"
-    local eat = "first"
-
-    local choices = {}
-    local buttons = {}
-
-    local function removeChoices()
-        for i = 1, 2 do
-            if choices[i] then
-                display.remove(choices[i])
-                choices[i] = nil
-            end
-            if buttons[i] then
-                display.remove(buttons[i])
-                buttons[i] = nil
-            end
-        end
-    end
-
-    local function EatTapped()
-        removeChoices()
-		if bg then
-			display.remove(bg)  -- 기존 배경 제거
-		end
-		bg = display.newImage("Image/cutscene/cutscene_7.png")  -- 새로운 배경 생성
-		bg.x = display.contentCenterX  -- 배경의 x 좌표 설정
-		bg.y = display.contentCenterY  -- 배경의 y 좌표 설정
-		bg:toBack()  -- 배경을 뒤로 보이게 설정
-
-	    content.text = "이건 칠면조 다리가 아니잖아 이건, \n이건............!"
-		
-		timer.performWithDelay(1800, function()
-			composer.gotoScene("diag_outside")
-		end)	
-    end
-
-    local function NotEatTapped(event)
-        if choices[2] and event.target == choices[2] then
-            choices[2].text = "먹는다."
-            eat = "second"
-            choices[2]:removeEventListener("tap", NotEatTapped)
-            choices[2]:addEventListener("tap", EatTapped)
-        end
-    end
-
-    function nextScript()
+    function nextScript(event)
         index = index + 1
         if (index > #Data) then
-            composer.gotoScene("diag_outside")
-            composer.removeScene("diag_table")
+            composer.gotoScene("after_diag_table")
             return
-        end
-
-        if (Data[index].eat == eat) then
-            dialogueBox:removeEventListener("tap", nextScript)
-
-            choices[1] = display.newText("먹는다.", display.contentCenterX * 1.15, display.contentCenterY + 180, 300, 50)
-            choices[2] = display.newText("먹지 않는다.", display.contentCenterX * 1.15, display.contentCenterY + 230, 300, 50)
-            for i = 1, 2 do
-                choices[i]:setFillColor(1)
-                choices[i].size = 30
-            end
-
-            buttons[1] = display.newRect(display.contentCenterX * 1.05 - 22, display.contentCenterY + 175, 165, 45)
-            buttons[1]:setFillColor(0)
-            buttons[1].alpha = 0.5
-            buttons[2] = display.newRect(display.contentCenterX * 1.05 - 22, display.contentCenterY + 225, 165, 45)
-            buttons[2]:setFillColor(0)
-            buttons[2].alpha = 0.5
-
-            choices[1]:addEventListener("tap", EatTapped)
-            choices[2]:addEventListener("tap", NotEatTapped)
-        elseif (eat == "second") then
-            removeChoices()
         end
 
         content.text = Data[index].dialogue or ""
